Let HikariCP handle the MySQL connection(s)
Use of MySQL is still discouraged, because these changes don't solve the issues where some data can reset in some situations.

The purpose of this commit is to delegate the responsibility of connection pooling, and to resolve the problem of retrying indefinately leading to hard to debug problems with bad configuration.